TAStudio: refactored ready for GDI+ implementaton
